some additional spike

+++ bin/kill.sh
@@ -1,2 +1,2 @@
 ./list.sh | awk '{print($2)}' | xargs kill -9
-find ./tmp -type s -delete
+[ -d ./tmp ] && find ./tmp -type s -delete
